+++ views/graph.py
@@ -21,6 +21,30 @@ from ..datasource.netcdf import (
     CUMULATIVE_AGGREGATION_UNITS)
 
 
+def parse_aggvarname(aggvarname):
+    """Parse a combined agg var name.
+
+    >>> parse_aggvarname('s1_max')
+    ('s1', 'max')
+    >>> parse_aggvarname('s1_cum_negative')
+    ('s1', 'cum_negative')
+    >>> parse_aggvarname('infiltration_rate_cum_positive')
+    ('infiltration_rate', 'cum_positive')
+    """
+    # Aggregation methods unfortunately can contain underscores; for now only
+    # these two cases are known.
+    # TODO: improve this, e.g., make more generic, because extra cases will
+    # need to be added later
+    if (aggvarname.endswith('cum_positive') or
+            aggvarname.endswith('cum_negative')):
+        varname, agg_method, sign = aggvarname.rsplit('_', 2)
+        return varname, "_".join([agg_method, sign])
+
+    # Works only for aggregation methods without underscores
+    varname, agg_method = aggvarname.rsplit('_', 1)  # maxsplit = 1
+    return varname, agg_method
+
+
 def generate_parameter_config(subgrid_map_vars, agg_vars):
     """Dynamically create the parameter config.
 
@@ -37,8 +61,11 @@ def generate_parameter_config(subgrid_map_vars, agg_vars):
     verbose_agg_method = {
         'min': 'minimum',
         'max': 'maximum',
-        'cum': 'cumulative',
+        'cum': 'net cumulative',
         'avg': 'average',
+        'med': 'median',
+        'cum_positive': 'positive cumulative',
+        'cum_negative': 'negative cumulative',
         }
 
     for varname in subgrid_map_vars:
@@ -51,17 +78,22 @@ def generate_parameter_config(subgrid_map_vars, agg_vars):
             config['h'].append(d)
 
     for aggvarname in agg_vars:
-        _varname, _agg_method = aggvarname.rsplit('_', 1)  # maxsplit = 1
+        _varname, _agg_method = parse_aggvarname(aggvarname)
         varinfo = agg_vars_mapping[_varname]
-        agg_method = verbose_agg_method[_agg_method]
+        try:
+            agg_method_display_name = verbose_agg_method[_agg_method]
+        except KeyError:
+            log("Unknown agg method: %s" % _agg_method, level='CRITICAL')
+            agg_method_display_name = _agg_method
 
         # Adjust the unit for cumulative method
-        if _agg_method == 'cum':
+        if _agg_method.startswith('cum'):
             unit = CUMULATIVE_AGGREGATION_UNITS[_varname]
         else:
             unit = varinfo[1]
 
-        d = {'name': '%s %s' % (agg_method.capitalize(), varinfo[0]),
+        d = {'name': '%s %s' % (agg_method_display_name.capitalize(),
+                                varinfo[0]),
              'unit': unit, 'parameters': aggvarname}
         if _varname in Q_TYPES:
             config['q'].append(d)