+++ CHANGES.rst
@@ -5,6 +5,8 @@ threedi-qgis-plugin changelog
 0.8.3 (unreleased)
 ------------------
 
+- Update cumulative aggregation methods.
+
 - The user_ref field now has the following format:
   ``<content>.code#<content>.id#<table_name>#calc_pnt_nr``